requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgi-cors==1.0.1
asgiref==3.9.1
billiard==4.2.1
botocore==1.40.24
celery==5.5.3
channels==4.3.1
channels_redis==4.3.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.4
django-filter==25.1
django-ninja==1.4.3
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
h11==0.16.0
iniconfig==2.1.0
jmespath==1.0.1
kombu==5.5.4
msgpack==1.1.1
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
redis==6.4.0
six==1.17.0
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
vine==5.1.0
wcwidth==0.2.13
websockets==15.0.1
